Expose MetadataBuilder, assign_paths(), insert_paths()
