+++ include/gbwtgraph/internal.h
@@ -94,103 +94,6 @@ struct ManualTSVWriter
 //------------------------------------------------------------------------------
 
 /*
-  FIXME: This should be made public.
-
-  A structure for building GBWT metadata.
-
-  Constructor and the methods for handling paths/walks throw `std::runtime_error`
-  on failure.
-*/
-struct MetadataBuilder
-{
-  constexpr static size_t NO_FIELD = std::numeric_limits<size_t>::max();
-  struct PathMetadataBuilder
-  {
-    std::regex parser;
-
-    // Mapping from regex submatches to GBWT path name components.
-    size_t sample_field, contig_field, haplotype_field, fragment_field;
-
-    PathSense sense;
-
-    PathMetadataBuilder(const std::string& path_name_regex, const std::string& path_name_prefix, PathSense path_sense);
-  };
-
-  std::vector<PathMetadataBuilder> path_name_formats;
-
-  // GBWT metadata.
-  std::map<std::string, size_t> sample_names, contig_names;
-  std::set<std::pair<size_t, size_t>> haplotypes; // (sample id, phase id)
-  std::vector<std::vector<gbwt::PathName>> path_names;
-  std::map<gbwt::PathName, size_t> counts;
-
-  bool ref_path_sample_warning;
-
-  // Construct a MetadataBuilder with no path name formats.
-  MetadataBuilder();
-  
-  // Construct a MetadataBuilder from pre-existing metadata and no path name formats.
-  MetadataBuilder(const gbwt::Metadata& source);
-
-  // Construct a MetadataBuilder with one path name format.
-  MetadataBuilder(const std::string& path_name_regex, const std::string& path_name_prefix, PathSense path_sense = PathSense::GENERIC);
-
-  // Register a format for parsing path names. Formats are tried in order until one matches.
-  void add_path_name_format(const std::string& path_name_regex, const std::string& path_name_prefix, PathSense path_sense);
-  
-  // Add a path defined by libhandlegraph metadata to the given job.
-  // Doesn't create metadata for samples or contigs if the no-name sentinel is
-  // passed for a sense that usually has them.
-  void add_path(PathSense sense, const std::string& sample_name, const std::string& locus_name, size_t haplotype, size_t phase_block, const handlegraph::subrange_t& subrange, size_t job = 0);
-  
-  // Parse a path name using a regex to determine metadata, and assign it to the given job.
-  void add_path(const std::string& name, size_t job = 0);
-
-  // Add a path based on walk metadata and assign it to the given job.
-  void add_walk(const std::string& sample, const std::string& haplotype, const std::string& contig, const std::string& start, size_t job = 0);
-
-  // Add a haplotype path and assign it to the given job.
-  void add_haplotype(const std::string& sample, const std::string& contig, size_t haplotype, size_t fragment, size_t job = 0);
-
-  // Add a named path as a generic named path and assign it to the given job.
-  void add_generic_path(const std::string& name, size_t job = 0);
-  
-  bool empty() const { return this->path_names.empty(); }
-
-  // Build GBWT metadata from the current contents.
-  // Paths come out ordered by job.
-  gbwt::Metadata get_metadata() const;
-
-  void clear()
-  {
-    this->sample_names = std::map<std::string, size_t>();
-    this->contig_names = std::map<std::string, size_t>();
-    this->haplotypes = std::set<std::pair<size_t, size_t>>();
-    this->path_names = std::vector<std::vector<gbwt::PathName>>();
-    this->counts = std::map<gbwt::PathName, size_t>();
-  }
-
-  static std::vector<std::string> map_to_vector(const std::map<std::string, size_t>& source)
-  {
-    std::vector<std::string> result(source.size());
-    for(auto& name : source) { result[name.second] = name.first; }
-    return result;
-  }
-
-private:
-  void add_path_name(const gbwt::PathName& path_name, size_t job)
-  {
-    if(job >= this->path_names.size())
-    {
-      this->path_names.resize(job + 1);
-    }
-    this->path_names[job].push_back(path_name);
-  }
-};
-
-//------------------------------------------------------------------------------
-
-/*
   A HandleGraph implementation that stores graph topology without sequences.
 */
 class EmptyGraph : public HandleGraph
